Bus basically free version

+++ constant.h
@@ -68,7 +68,7 @@ const float MIN_LENGTH_SECS = 5.0; //secs
 
 
 const float MATCHING_THRESHOLD = 0.15;
-const float SEGMENT_LENGTH = 40;
+const int SEGMENT_LENGTH = 40;
 
 #endif
